requirements.txt
streamlit
crewai
python-dotenv
pillow
google-generativeai==0.5.4
google-ai-generativelanguage==0.6.4
protobuf==4.25.3


